Updated data visualizer and move Projection to new file while updating
