+++ yellowbrick/features/base.py
@@ -20,16 +20,13 @@ Base classes for feature visualizers and feature selection tools.
 
 import numpy as np
 import matplotlib as mpl
-import matplotlib.pyplot as plt
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 import warnings
 from enum import Enum
 
 from yellowbrick.base import Visualizer
 from yellowbrick.utils import is_dataframe
 from yellowbrick.style import resolve_colors
-from yellowbrick.draw import manual_legend
-from yellowbrick.exceptions import YellowbrickValueError, YellowbrickWarning,  NotFitted
+from yellowbrick.exceptions import YellowbrickValueError, YellowbrickWarning
 
 from sklearn.base import TransformerMixin
 
@@ -279,10 +276,8 @@ class DataVisualizer(MultiFeatureVisualizer):
             raise YellowbrickValueError(
                     "unknown target color type '{}'".format(self._target_color_type))
 
-        # Draw the instances
-        self.draw(X, y, **kwargs)
-
         # Fit always returns self.
+        # NOTE: cannot call draw in fit to support data transformers
         return self
 
     def _determine_target_color_type(self, y):
@@ -312,130 +307,3 @@ class DataVisualizer(MultiFeatureVisualizer):
                 "could not determine target color type "
                 "from target='{}' to '{}'"
             ).format(self.target_type, self._target_color_type))
-
-
-##########################################################################
-## Projection Visualizers
-##########################################################################
-
-class ProjectionVisualizer(DataVisualizer):
-    
-    def __init__(self, transformer, ax=None, features=None, classes=None, color=None,
-             colormap=None, target_type="auto", proj_dim=2, alpha=0.75, **kwargs):
-        super(ProjectionVisualizer, self).__init__(ax=ax, features=features, 
-                                                     classes=classes, color=color,
-                                                     colormap=colormap, 
-                                                     target_type=target_type, **kwargs)
-        self.transformer=transformer
-        if proj_dim not in [2, 3]:
-            raise YellowbrickValueError("Projection dimensions must be either 2 or 3")
-        self.proj_dim = proj_dim
-        self.alpha = alpha
-        self._cax = None
-
-    @property
-    def cax(self):
-        """
-        The axes of the colorbar, right of the scatterplot.
-        """
-        if self._cax is None:
-            raise AttributeError(
-                "this visualizer does not have a histogram for the X axis"
-            )
-        return self._cax       
-        
-    def _layout_colorbar(self):
-        """
-        Creates the layout for colorbar. The colorbar is added to the right of the 
-        scatterplot 
-        """
-        # Ensure matplotlib version compatibility
-        if make_axes_locatable is None:
-            raise YellowbrickValueError((
-                "heatmap requires matplotlib 2.0.2 or greater "
-                "please upgrade matplotlib or set heatmap=False on the visualizer"
-            ))
-
-        # Create the new axes for the colorbar
-        divider = make_axes_locatable(self.ax)
-        self._cax = divider.append_axes("right", size="5%", pad=0.3)
-        
-    def fit(self, X, y=None, **kwargs):
-        
-        self.fit_transform(X, y, **kwargs)
-        return self
-        
-    def fit_transform(self, X, y=None, **kwargs):
-
-        # Transformer will be  defined either in the subclasses or explicitly
-        self.transformed_features_ = self.transformer.fit_transform(X)
-        super(ProjectionVisualizer, self).fit(X, y, **kwargs)
-        return self.transformed_features_    
-        
-    def transform(self, X, y=None, **kwargs):
-        try:
-            return self.transformer.transform(X)
-        except AttributeError as e:
-            raise AttributeError(str(e) + " try using fit_transform instead.")
-
-    # Not sure about the function signature      
-    def draw(self, X=None, y=None):
-        X = self.transformed_features_
-        scatter_kwargs = self._determine_scatter_kwargs(y);
-        
-        if self.proj_dim == 2:
-            self._scatter = self.ax.scatter(X[:,0], X[:,1], **scatter_kwargs)
-
-        if self.proj_dim == 3:
-            self.fig = plt.figure()
-            self.ax = self.fig.add_subplot(111, projection='3d')
-            self._scatter = self.ax.scatter(X[:, 0], X[:, 1], X[:, 2], **scatter_kwargs)
-        
-        return self.ax
-
-    def finalize(self):
-        
-        if self._target_color_type == TargetType.DISCRETE:
-            # Add the legend
-            manual_legend(self, self.classes_, list(self._colors.values()),
-                          frameon=True)
-
-        elif self._target_color_type == TargetType.CONTINUOUS:
-            if(self.proj_dim==3):
-                self.fig.colorbar(self._scatter, ax=self.ax)
-            
-            else:
-                self._layout_colorbar()
-                # Manually draw the colorbar.
-                self.cbar = mpl.colorbar.ColorbarBase(self.cax, cmap=self._colors, 
-                                                  norm=self._norm)
-
-    def _determine_scatter_kwargs(self, y=None):
-        scatter_kwargs = {"alpha": self.alpha}
-        # Determine the colors
-        if self._target_color_type == TargetType.SINGLE:
-            scatter_kwargs["c"] = self._colors
-
-        elif self._target_color_type == TargetType.DISCRETE:
-            if y is None:
-                raise YellowbrickValueError("y is required for discrete target")
-
-            try:
-                scatter_kwargs["c"] = [self._colors[self.classes_[yi]] for yi in y]
-            except IndexError:
-                raise YellowbrickValueError("Target needs to be label encoded.")
-
-
-        elif self._target_color_type == TargetType.CONTINUOUS:
-            if y is None:
-                raise YellowbrickValueError("y is required for continuous target")
-
-            scatter_kwargs["c"] = y
-            scatter_kwargs["cmap"] = self._colors
-            self._norm = mpl.colors.Normalize(vmin=self.range_[0], vmax=self.range_[1])
-
-        else:
-            # Technically this should never be raised
-            raise NotFitted("could not determine target color type")
-                # Draw the scatter plot with the associated colors and alpha
-        return scatter_kwargs;
